unit tests from cosmetic_filter_cache.rs moved to tests/unit/cosmetic_filter_cache.rs
